ship variant 3
